Move restaurnt components in root component folder. Rename components according to strategy entity-typeOfComponent
